feat: vault backup banner
